Koz reported that Noise Removal wasn't accepting typed values properly and this should fix that.

+++ sf-cvs/trunk/audacity-src/src/effects/NoiseRemoval.h
@@ -172,9 +172,9 @@ private:
    wxTextCtrl *mFreqT;
    wxTextCtrl *mTimeT;
 
-   float       mGain;
-   float       mFreq;
-   float       mTime;
+   double      mGain;
+   double      mFreq;
+   double      mTime;
 
 private:
    DECLARE_EVENT_TABLE()
